--- generators/app/templates/infrastructure/src/startup/subscriptions/middleware/correlation.js
@@ -1,7 +1,0 @@
-const { correlationManager } = require("../../../correlation");
-
-const correlation = async (ctx, next) => {
-    return await correlationManager.useCorrelationId(ctx.message.headers?.pubSubCorrelationId, next)
-};
-
-module.exports = { correlation };
