+++ generators/app/templates/infrastructure/src/subscriptions/index.js
