+++ generators/app/templates/infrastructure/src/subscriptions/middleware/index.js
